Rename bot to seabird and simplify
