+++ seabird/bot.go
@@ -1,0 +1,236 @@
+package seabird
+
+import (
+	"crypto/tls"
+	"errors"
+	"fmt"
+	"io"
+	"log"
+	"net"
+
+	"github.com/BurntSushi/toml"
+
+	"github.com/belak/go-plugin"
+	"github.com/belak/go-seabird/internal"
+	"github.com/belak/irc"
+)
+
+type coreConfig struct {
+	Nick string
+	User string
+	Name string
+	Pass string
+
+	Host        string
+	TLS         bool
+	TLSNoVerify bool
+
+	Cmds   []string
+	Prefix string
+
+	Plugins []string
+
+	Quiet bool
+}
+
+// A Bot is our wrapper around the irc.Client. It could be used for a general
+// client, but the provided convenience functions are designed around using this
+// package to write a bot.
+type Bot struct {
+	mux *BasicMux
+
+	// Config stuff
+	confValues map[string]toml.Primitive
+	md         toml.MetaData
+	config     coreConfig
+
+	// Internal things
+	client   *irc.Client
+	registry *plugin.Registry
+}
+
+// NewBot will return a new Bot given the name of a toml config file.
+func NewBot(confReader io.Reader) (*Bot, error) {
+	var err error
+
+	b := &Bot{
+		NewBasicMux(),
+		make(map[string]toml.Primitive),
+		toml.MetaData{},
+		coreConfig{},
+		nil,
+		plugins.Copy(),
+	}
+
+	// Decode the file, but leave all the config sections intact so we can
+	// decode those later.
+	b.md, err = toml.DecodeReader(confReader, &b.confValues)
+	if err != nil {
+		return nil, err
+	}
+
+	// Load up the core config
+	err = b.Config("core", &b.config)
+	if err != nil {
+		return nil, err
+	}
+
+	commandMux := NewCommandMux(b.config.Prefix)
+	mentionMux := NewMentionMux()
+
+	b.mux.Event("PRIVMSG", commandMux.HandleEvent)
+	b.mux.Event("PRIVMSG", mentionMux.HandleEvent)
+
+	// Register all the things we want with the plugin registry.
+	b.registry.RegisterProvider(func() (*Bot, *BasicMux, *CommandMux, *MentionMux) {
+		return b, b.mux, commandMux, mentionMux
+	})
+
+	return b, nil
+}
+
+// CurrentNick returns the current nick of the bot.
+func (b *Bot) CurrentNick() string {
+	return b.client.CurrentNick()
+}
+
+// Config will decode the config section for the given name into the given
+// interface{}
+func (b *Bot) Config(name string, c interface{}) error {
+	if v, ok := b.confValues[name]; ok {
+		return b.md.PrimitiveDecode(v, c)
+	}
+
+	return fmt.Errorf("Config section for %q missing", name)
+}
+
+// Send is a simple function to send an IRC event
+func (b *Bot) Send(m *irc.Message) {
+	b.client.WriteMessage(m)
+}
+
+// Reply to an irc.Message with a convenience wrapper around fmt.Sprintf
+func (b *Bot) Reply(m *irc.Message, format string, v ...interface{}) error {
+	if len(m.Params) < 1 || len(m.Params[0]) < 1 {
+		return errors.New("Invalid IRC message")
+	}
+
+	target := m.Prefix.Name
+	if m.FromChannel() {
+		target = m.Params[0]
+	}
+
+	b.Send(&irc.Message{
+		Prefix:  &irc.Prefix{},
+		Command: "PRIVMSG",
+		Params: []string{
+			target,
+			fmt.Sprintf(format, v...),
+		},
+	})
+
+	return nil
+}
+
+// MentionReply acts the same as Bot.Reply but it will prefix it with the user's
+// nick if we are in a channel.
+func (b *Bot) MentionReply(m *irc.Message, format string, v ...interface{}) error {
+	if m.FromChannel() {
+		format = "%s: " + format
+		v = internal.Prepend(v, m.Prefix.Name)
+	}
+
+	return b.Reply(m, format, v...)
+}
+
+// CTCPReply is a convenience function to respond to CTCP requests.
+func (b *Bot) CTCPReply(m *irc.Message, format string, v ...interface{}) error {
+	if m.Command != "CTCP" {
+		return errors.New("Invalid CTCP message")
+	}
+
+	b.Send(&irc.Message{
+		Prefix:  &irc.Prefix{},
+		Command: "NOTICE",
+		Params: []string{
+			m.Prefix.Name,
+			fmt.Sprintf(format, v...),
+		},
+	})
+
+	return nil
+}
+
+func (b *Bot) handshake() {
+	b.client.Writef("CAP END")
+	b.client.Writef("NICK %s", b.config.Nick)
+	b.client.Writef("USER %s 0.0.0.0 0.0.0.0 :%s", b.config.User, b.config.Name)
+}
+
+// Write will write an raw IRC message to the stream
+func (b *Bot) Write(line string) {
+	b.client.Write(line)
+}
+
+// Writef is a convenience method around fmt.Sprintf and Bot.Write
+func (b *Bot) Writef(format string, args ...interface{}) {
+	b.client.Writef(format, args...)
+}
+
+func (b *Bot) handler(c *irc.Client, m *irc.Message) {
+	if m.Command == "001" {
+		log.Println("Connected")
+
+		for _, v := range b.config.Cmds {
+			b.Write(v)
+		}
+	}
+
+	b.mux.HandleEvent(b, m)
+}
+
+// Run starts the bot and loops until it dies
+func (b *Bot) Run() error {
+	// TODO: We currently ignore the injector, but it could be nice to keep it
+	// around for optional plugins.
+	_, err := b.registry.Load(b.config.Plugins, nil)
+	if err != nil {
+		return err
+	}
+
+	// The ReadWriteCloser will contain either a *net.Conn or *tls.Conn
+	var c io.ReadWriteCloser
+	if b.config.TLS {
+		conf := &tls.Config{
+			InsecureSkipVerify: b.config.TLSNoVerify,
+		}
+
+		c, err = tls.Dial("tcp", b.config.Host, conf)
+	} else {
+		c, err = net.Dial("tcp", b.config.Host)
+	}
+
+	if err != nil {
+		return err
+	}
+
+	// Create a client from the connection we've just opened
+	rc := irc.ClientConfig{
+		Nick: b.config.Nick,
+		Pass: b.config.Pass,
+		User: b.config.User,
+		Name: b.config.Name,
+
+		Handler: irc.HandlerFunc(b.handler),
+	}
+	b.client = irc.NewClient(c, rc)
+
+	b.client.DebugCallback = func(line string) {
+		if !b.config.Quiet {
+			log.Println(line)
+		}
+	}
+
+	// Start the main loop
+	return b.client.Run()
+}
