+++ seabird/mux_command.go
@@ -1,4 +1,4 @@
-package bot
+package seabird
 
 import (
 	"sort"
