+++ seabird/mux_mention.go
@@ -1,4 +1,4 @@
-package bot
+package seabird
 
 import (
 	"strings"
